refactor: encrypt 추상화
