+++ api-member/src/main/java/com/style/member/infra/encrypt/PasswordEncoder.java
@@ -1,0 +1,9 @@
+package com.style.member.infra.encrypt;
+
+public interface PasswordEncoder {
+
+    String encode(String password);
+
+    boolean verifyPassword(String requestPassword, String memberPassword);
+
+}
