+++ api-member/src/test/java/com/style/fixture/MemberFixture.java
@@ -1,7 +1,6 @@
 package com.style.fixture;
 
 import com.style.auth.presentation.request.SignInRequest;
-import com.style.member.presentation.request.SignOffRequest;
 import com.style.member.presentation.request.SignUpRequest;
 
 public class MemberFixture {
